refactor: Move decoderEVM to its own file.
